+++ src/Ext/WarheadType/Detonate.cpp
@@ -419,17 +419,11 @@ void WarheadTypeExt::ApplyShieldModifiers(TechnoClass* pTarget)
 				if (this->Shield_ReplaceOnly && this->Shield_InheritStateOnReplace)
 				{
 					pShield->SetHP((int)(shieldType->Strength * ratio));
-
-					if (this->Shield_ReplaceOnly && this->Shield_InheritStateOnReplace)
-					{
-						pShield->SetHP((int)(shieldType->Strength * ratio));
-
-						if (pShield->GetHP() == 0)
-						{
-							pShield->SetRespawn(shieldType->Respawn_Rate, shieldType->Respawn, shieldType->Respawn_Rate,
-								shieldType->Respawn_RestartInCombat, -1, true, shieldType->Respawn_Anim);
-						}
-					}
+					// Value doesn't matter here, it's just for restarting timer
+					pShield->SetRespawn(0, shieldType->Respawn, shieldType->Respawn_Rate,
+						shieldType->Respawn_RestartInCombat, -1, true, shieldType->Respawn_Anim);
+					pShield->SetSelfHealing(0, shieldType->SelfHealing, shieldType->SelfHealing_Rate,
+						shieldType->SelfHealing_RestartInCombat, -1, true);
 				}
 			}
 		}